Initial version
The only working player is oto-v2. But it works pretty bad. I have
serious issue with audio streams:
- sometimes there is no response after starting the program
  it takes up to 10 restarts to make it work
- but even when it works, it stops responding after 5-10 seconds
- the response voice is very low and slow

I suppose there are issues with the sample rates. People on OpenAI
community forum mentioned that OpenAI uses 24kHz sample rate for input
and output but I could not find any mention of that in the docs.